database file for alx bookstore

+++ alx_book_store.sql
@@ -32,7 +32,7 @@ CREATE TABLE Order_Details (
 order_detail_id INT AUTO_INCREMENT PRIMARY KEY,
 order_id INT,
 book_id INT,
-quantity DOUBLE,
+quantity DOUBLE(10,2),
 FOREIGN KEY (order_id) REFERENCES Orders(order_id),
 FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
